♻️ control-plane: move ClusterTenant CR bridge out of routes into core/cluster-tenants
The DB→K8s ClusterTenant CR projection was misfiled under routes/internal/
(reserved for auth-less HTTP endpoints) despite being pure domain logic. Move
it to core/cluster-tenants/cr-bridge.ts, next to the provisioner it feeds, so
routes stay thin HTTP handlers.

Also relocate the shared CRD constants from routes/internal/crd-constants.ts to
a neutral shared/crd-constants.ts (mirroring the operator's shared/ layout) so
core no longer imports from routes — fixing the backwards layer dependency the
move would otherwise create. Pure file moves + import rewires; no behaviour
change. control-plane: tsc clean, 408 tests pass.
