+++ apps/control-plane/src/routes/internal/projection-repair.ts
@@ -1,7 +1,7 @@
 import type * as k8s from "@kubernetes/client-node";
 import type { PrismaClient } from "@prisma/client";
 
-import { OPENCRANE_API_GROUP, OPENCRANE_API_VERSION, TENANT_CRD_PLURAL, POLICY_CRD_PLURAL } from "./crd-constants.js";
+import { OPENCRANE_API_GROUP, OPENCRANE_API_VERSION, TENANT_CRD_PLURAL, POLICY_CRD_PLURAL } from "../../shared/crd-constants.js";
 import type { ProjectionRepairEntry, ProjectionRepairReport } from "./projection-repair.types.js";
 
 /**
